Neptune controller (#319)
* more work

* more work

* more work, gyro

* more work

* more work

* hotfixes

* fix a few things

* I believe I have gyro angles right, accelero is crippled

* taking a break @CasperH2O have fun

* fixed values but accel angles are still incorrect

* fix accelero

* prevent crash if Neptune Controller can't be initialized

* fix XInput controller disconnect

* greatly improve the way we manage windows and child processes

* increase process mainwindow delay

* a bit more work on SensorFamily

* more work on neptune/jupiter

* oups

* Update ControllerService.cs

* code cleanup

* wip

* wip, implementing touchpad support

* oups

* oups

* done

* better

* improve

* stick by the book

* Update DS4Touch.cs

* do not store inputs in pipe buffer

* build 0.14.0.0

* fix trigger update from quicktools

* Update README.md

* SteamDeck 3D model for overlay.

* add missing line to use steamdeck model

* model highlight color to use AccentButtonBackground

* update libraries

* implement backup strategy on HidHide

* more work on log and a few fixes

* improved log

* deprecated

* more reliable UWP process handling

* trying to imporve FriendlyName capture

* Neptune controller vibration strength.

* a few edits

* my fault

* fix device rumble and add rumble on vibration strength changes

* more work on FriendlyName

* misc

* use SymLink rather than InstanceId

* few fixes

* add missing SetVibrationStrength to NeptuneController

* even better

* lot of work on making Symlink handling more robust

* improve steam deck rumble experience

* increase a few timers

* migrate a few vars

* move ControllerInput

* quick cleanup

* migrate a few libraries to core

- HelixToolkit.Core.Wpf
- LiveCharts.Wpf.Core

* StopListening() on RefreshSensors()

* build 0.14.0.1

* qol

* Steamdeck APU TDP and Gfx clock range improvements.

* Improve Steamdeck 3D Model trigger rotation point

* Fix Steamdeck 3D model motor position

* Improve Steamdeck 3D model colors.

* use TieredPGO

* implement Lizard Mode support (both controller page and hotkeys)

* update quicktools layout on show() or hide()

* build 0.14.0.2

* implementing IPlatform class and SteamPlatform support

* better way to handle platform detection

* adopting (is null) and (is not null) thanks to C# 9.0

* code cleanup

* build 0.14.0.3

* prevent a race condition with PlatformManager

* update log on unsupported DInputController (Xbox360, XboxOne)

* Renaming Path to InstallPath across a few functions to prevent ambiguity

* prevent ProcessManager hook from being active before manager is started

* prevent a crash on non existing process

* fix an issue with controller hide/reveal when using OEM special key

* reduce cpu overhead

* further reduce cpu overhead

* tons of work

* more work

* fix known issue with quick profile page on profile swap

* release lock

* further improve profile swapping and self check

* improve controller page UI

* implement SettingsManager, SetProperty force event raise

* add missing DrawControls

* revert profile change check

* fix remaining profile issues

* moving antideadzone settings away from UMC

* improved readability of the controller page

* build 0.14.0.4

* tweaking logs

* prevent InputsManager from messing with inputs on lockscreen

* build 0.14.0.5

* bonus keys for steamdeck owners

* more work on SD model

* qol

* fix quick profile Right anti-deadzone slider

* explorer.exe to return Ignored process filter

* prevent CPU overhead when tweaking profile settings on quick profiles page

* code cleanup

* qol

* build 0.14.0.6

* move controller connect/lookup logic to ControllerManager

* prevent issue with rumble when CPU is busy

* simplify loop

* Fix anti deadzone function and move applying to thumbsticks prior to UMC

* prevent crash when parsing outdated hotkey

* small hack to prevent issues when an unused hotkey is pressed while another one is already pressed

* Improve circularity function

* Add ImproveCircularity in more UMC locations.

* Restore UMC Anti Deadzone with adjusted naming.

* prevent useless HidHide calls

* force updatelayout on quicktools visibility state change

* fix an issue where UI would hang due to a profile update loop

* refactoring a few managers

* comments

* implement screen resolution and frequency dropdowns on quicktools

* clear resolution list on hardware changes

* hotfix resolution/frequency change crash

* prevent button round issue on controller list

* prevent crash when trying to manipulate HidHide if a Client is already open

* misc

* improve service stability on system wakeup

* fix an issue with PowerManager IsSessionLocked check when running as a service

* build 0.14.0.7

* compile ModernWpf to use .net7

* sort resolutions and frequencies quicktools dropdowns

* work on PowerStatus indicator (wip)

* minor UI edit

* quicktools to call UpdateDefaultStyle on visibility change

* add BatteryIndicatorLifeRemaining

* build 0.14.0.8

* improve ControllerManager reliability

* remove inpoutx64.dll, was causing EasyAntiCheat to trigger

sadly, doing so, we're losing the ability to monitor system values

* Update HandheldCompanion.csproj

* improve BatteryIndicatorLifeRemaining

* improve battery reading

* improve battery reading

* code cleanup

+++ ControllerCommon/HidHide.cs
@@ -5,58 +5,127 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
-using static PInvoke.Kernel32;
+using System.Linq;
 
 namespace ControllerCommon
 {
     public static class HidHide
     {
-        private static HidHideControlService service;
+        private static Process process;
         static HidHide()
         {
-            service = new HidHideControlService();
+            var service = new HidHideControlService();
 
             if (!service.IsInstalled)
             {
                 LogManager.LogCritical("HidHide is missing. Please get it from: {0}", "https://github.com/ViGEm/HidHide/releases");
                 throw new InvalidOperationException();
             }
+
+            // verifying HidHide is installed
+            string InstallPath = RegistryUtils.GetHKLM(@"SOFTWARE\Nefarius Software Solutions e.U.\HidHide", "Path");
+            if (!string.IsNullOrEmpty(InstallPath))
+                InstallPath = Path.Combine(InstallPath, "x64", "HidHideCLI.exe");
+
+            process = new Process
+            {
+                StartInfo =
+                {
+                    UseShellExecute = false,
+                    RedirectStandardOutput = true,
+                    RedirectStandardError = true,
+                    CreateNoWindow = true,
+                    FileName = InstallPath,
+                    Verb = "runas"
+                }
+            };
         }
 
-        public static IReadOnlyList<string> GetRegisteredApplications()
+        public static List<string> GetRegisteredApplications()
         {
-            return service.ApplicationPaths;
+            try
+            {
+                var service = new HidHideControlService();
+                return service.ApplicationPaths.ToList();
+            }
+            catch { }
+
+            return new();
         }
 
-        public static IReadOnlyList<string> GetRegisteredDevices()
+        public static List<string> GetRegisteredDevices()
         {
-            return service.BlockedInstanceIds;
+            try
+            {
+                var service = new HidHideControlService();
+                return service.BlockedInstanceIds.ToList();
+            }
+            catch { }
+
+            return new();
+        }
+
+        public static bool IsRegistered(string InstanceId)
+        {
+            try
+            {
+                var registered = GetRegisteredDevices();
+                return registered.Contains(InstanceId);
+            }
+            catch { }
+
+            return false;
         }
 
         public static void UnregisterApplication(string fileName)
         {
             try
             {
-                service.RemoveApplicationPath(fileName);
+                var service = new HidHideControlService();
+                if (service.ApplicationPaths.Contains(fileName))
+                {
+                    service.RemoveApplicationPath(fileName);
+                    LogManager.LogInformation("HideDevice RemoveApplicationPath: {0}", fileName);
+                }
+            }
+            catch
+            {
+                process.StartInfo.Arguments = $"--app-unreg \"{fileName}\"";
+                process.Start();
+                process.WaitForExit();
+                process.StandardOutput.ReadToEnd();
+
                 LogManager.LogInformation("HideDevice RemoveApplicationPath: {0}", fileName);
             }
-            catch { }
         }
 
         public static void RegisterApplication(string fileName)
         {
             try
             {
-                service.AddApplicationPath(fileName);
+                var service = new HidHideControlService();
+                if (!service.ApplicationPaths.Contains(fileName))
+                {
+                    service.AddApplicationPath(fileName);
+                    LogManager.LogInformation("HideDevice AddApplicationPath: {0}", fileName);
+                }
+            }
+            catch
+            {
+                process.StartInfo.Arguments = $"--app-reg \"{fileName}\"";
+                process.Start();
+                process.WaitForExit();
+                process.StandardOutput.ReadToEnd();
+
                 LogManager.LogInformation("HideDevice AddApplicationPath: {0}", fileName);
             }
-            catch { }
         }
 
         public static void SetCloaking(bool status)
         {
             try
             {
+                var service = new HidHideControlService();
                 service.IsActive = status;
                 LogManager.LogInformation("HideDevice IsActive: {0}", status);
             }
@@ -67,8 +136,12 @@ namespace ControllerCommon
         {
             try
             {
-                service.RemoveBlockedInstanceId(deviceInstancePath);
-                LogManager.LogInformation("HideDevice RemoveBlockedInstanceId: {0}", deviceInstancePath);
+                var service = new HidHideControlService();
+                if (service.BlockedInstanceIds.Contains(deviceInstancePath))
+                {
+                    service.RemoveBlockedInstanceId(deviceInstancePath);
+                    LogManager.LogInformation("HideDevice RemoveBlockedInstanceId: {0}", deviceInstancePath);
+                }
             }
             catch { }
         }
@@ -77,8 +150,12 @@ namespace ControllerCommon
         {
             try
             {
-                service.AddBlockedInstanceId(deviceInstancePath);
-                LogManager.LogInformation("HideDevice AddBlockedInstanceId: {0}", deviceInstancePath);
+                var service = new HidHideControlService();
+                if (!service.BlockedInstanceIds.Contains(deviceInstancePath))
+                {
+                    service.AddBlockedInstanceId(deviceInstancePath);
+                    LogManager.LogInformation("HideDevice AddBlockedInstanceId: {0}", deviceInstancePath);
+                }
             }
             catch { }
         }
